add multi colour rules
